Improve my own customizations and some pages to support site.baseurl
In _config.yml site.url is the URL starting with http, and baseurl is the subdirectory following this url. For example, wwww.clean.com/example have www.clean.com as url and /example as baseurl. If you want to serve your site in a subfolder under the root URL then you need to set this in this way. The minimal-mistake template support this fully by referencing any links via {{ address | relative_url }} which effectively prepends the "address" by site.baseurl. However, my customized snippets and a few pages of mine did not support this and resolve to failure when setting my site under wwww2.coe.pku.edu.cn/faculty/benshuailyu/. This commit addresses this problem.