Standardize the output format to be that controlled by the global permlink setting in _config.yml
There is a central setting in _config.yml designatign the output format of the html pages. For example all the post are generated to be xxx/waves/ (i.e. generate for each post a folder under which there is a index.html). However because all the files under publication and teaching collections are copied from previous template where permlink is customized in each file front matter. Those permlink without .html extension does not work on PKU server although working on my PC and github. To eliminate all these troubles in this commit we remove all the customized permlinks in each file and rely on the global setting to generate the url, which now each file resolves into a folder.